refactoring all mutations and queries to APIClients folder
